New Assets handling class
If you really want to get obsessive about making sure everything is
cacheable then you need to take care of delivering all the files
yourself as at least Apache does not do a great job of setting all the
headers without a great deal of persuasion. This class delivers files
that are in the assets directory.