Array access
Add in a way of accessing individual array elements for get and post.
Does this in typical Php fashion by passing in an array with a path of
keys in it. There seems to be no standard function that lets you apply
an array of keys to a an array and get the result. But I may be wrong.